Extract spy to its own file
